fast_api/fastapi_basic/mysite4/schemas/tag.py
# schemas/tag.py

from pydantic import BaseModel, ConfigDict


class TagCreate(BaseModel):
    name: str


class TagResponse(BaseModel):
    id: int
    name: str

    model_config = ConfigDict(from_attributes=True)
