fast_api/fastapi_basic/pyproject.toml
[project]
name = "fastapi-basic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi-pagination>=0.15.10",
    "fastapi[standard]>=0.128.0",
    "openai>=2.20.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
